+++ website/lib/generated-meta.json
@@ -1,0 +1,8 @@
+{
+  "version": "0.9.1",
+  "mcpTools": 45,
+  "hooks": 12,
+  "restEndpoints": 112,
+  "testsPassing": 819,
+  "generatedAt": "2026-04-21T12:25:41.163Z"
+}
